+++ macros/stored_procedures/snowflake__stored_procedure.sql
@@ -1,6 +1,6 @@
 /*
   This materialization is used for creating stored procedure objects.
-  The idea behind this materialization is for ability to define CREATE STORED PROCEDURE statements and have dbt use the necessary logic
+  The idea behind this materialization is for ability to define create stored procedure statements and have dbt use the necessary logic
   of deploying the stored procedure in a consistent manner and logic.
   Adapted from https://github.com/venkatra/dbt_hacks
 
@@ -11,6 +11,8 @@
   {%- set identifier = config.get('override_name', default=model['alias'] ) -%}
   {%- set return_type = config.get('return_type', default='varchar' ) -%}
   {%- set execute_as = config.get('execute_as', default='owner' ) -%}
+  {%- set create_or_replace = config.get('create_or_replace', default=true) -%}
+  {%- set include_copy_grants = config.get('include_copy_grants', default=false) -%}
 
   {%- set target_relation = api.Relation.create( identifier=identifier, schema=schema, database=database) -%}
 
@@ -25,8 +27,20 @@
       --------------------------------------------------------------------------------------------------------------------
   -- build model
 
+  {% set copy_grants_statement = "" %}
+  {% if create_or_replace %}
+       {% set create_statement = "create or replace" %}
+       {% if include_copy_grants %}
+              {% set copy_grants_statement = "copy grants" %}
+       {% endif %}
+  {% else %}
+       {% set create_statement = "create or alter" %}
+  {% endif %}
+  
+       
+
   {% call statement('main') -%}
-    {{ dbt_dataengineers_materializations.snowflake_create_stored_procedure_statement(target_relation, preferred_language, parameters, return_type, execute_as, sql) }}
+    {{ dbt_dataengineers_materializations.snowflake_create_stored_procedure_statement(target_relation, create_statement, copy_grants_statement, preferred_language, parameters, return_type, execute_as, sql) }}
   {%- endcall %}
 
       --------------------------------------------------------------------------------------------------------------------